USER-MANAGEMENT/List-User/requirements.txt
Flask
Flask-Cors
python-dotenv
mysql-connector-python